mcp-servers/general/requirements.txt
sse-starlette
starlette
mcp
google-adk==1.8.0